Make j and k movements linewise for operators
This adds the same trick easymotion uses to make sure commands like
"d<keybindingForHopJ>" operate on entire lines instead of parts of
lines.

For example, if you had:
        v
This is the first line
and this is the second

And you tried to use the keybinding mentioned above, you would get:
This is  is the second

Whereas you would expect it to delete the entirety of both lines. This
commit changes hop so that HopJ and HopK act how you would expect.