Make motions inclusive/linewise as needed
This change makes most motions inclusive and a handful of motions
linewise. An inclusive motion includes the target character when
yanking, deleting, etc. This should align with expectations about how
the hop motions work. A linewise motion operates on entire lines.

This code checks that the user is in "no" mode which means this change
shouldn't do anything if the user was already in visual mode or had
explicitly set the motion to inclusive, exclusive, or linewise.