Fix truncated hints after unicode in wrap mode.
If some unicode / multi-byte is present at column `col` on a given line
and that `wrap` is set, then after `col`, some hints will be truncated.
This was implemented in the first place to hint _only_ what’s visible
(so that we don’t waste time hinting stuff the user cannot even see). In
the case of `wrap` mode, the computation of what’s visible was buggy.

This commit patches this situation by converting the width into a byte
width, allowing to take into unicode and more specifically, multi-byte
fucking characters. Should work with emojis, too.

Relates to #72.