Add HintDirection and read it from opts.
This first commit allows to apply the hints to only the previous part of
the cursor or what comes after.

This commit doesn’t currently check the column, only the line. The next
commit will add support for before / after on the current line and
should also add support for a way to reduce apply hints only to the
current line.

This is an alternative implementation of #59.

Relates to #71, #91.