Simplify create_jump_targets_for_line (context).
The context is not entirely needed here, so it’s better to just make
this function take several arguments. Also, for some other jump target
generators, where the context is not actually needed, but we would still
want to use that create_jump_targets_for_line function (for instance for
modes using only the current line), passing the context is a waste of
time. So better to do it that way.