fix: fix "col value outside range" error
getcurpos()'s curswant refers to the _preferred_ column, and it
signifies to what column the cursor should get back to after moving over
empty or shorter lines. It does not refer to the actual cursor position,
and cannot be used here because it can report a column that does not
actually exist on the current line.

virtcol() would be more correct, as it reports the actual column.
However, it counts hard tabs as their screen width, could still cause
out of range errors.

Instead get the current line up to the cursor and calculate the column
with strwidth().

Fixes #261