Fix column out of range for HopLine*.
This commit has the disavantage of preventing people from jumping to
empty lines if they have shifted their view. However, I think it’s okay
because we cannot place a hint in a buffer that is not visible, unless
we start hacking around with the text_column thing, and I don’t think
it’s worth it for now.

Fixes #292.