Fix catching user's initial config state
Hint state was taken out one layer upwards to main interfaces in order to fix
hop pattern as it adds virtual cursor which corrupts user settings.
Because of that, hint state has to be saved before that happens.

This is a direct follow up to #148. This rewrite should solve this issue
entirely for all cases.

Signed-off-by: afterlook <29754364+afterlook@users.noreply.github.com>